Move to middleware file
