+++ lib/crowi/express-init.js
@@ -8,38 +8,10 @@ module.exports = function(crowi, app) {
   var methodOverride = require('method-override')
   var passport = require('passport')
   var session = require('express-session')
-  var basicAuth = require('basic-auth-connect')
   var flash = require('connect-flash')
   var cons = require('consolidate')
-  var i18next = require('i18next')
-  var i18nFsBackend = require('i18next-node-fs-backend')
-  var i18nSprintf = require('i18next-sprintf-postprocessor')
-  var i18nMiddleware = require('i18next-express-middleware')
-  var i18nUserSettingDetector = require('../util/i18nUserSettingDetector')
   var env = crowi.node_env
-  const middlewares = require('../middlewares')(crowi, app)
-
-  var User = crowi.model('User')
-
-  var lngDetector = new i18nMiddleware.LanguageDetector()
-  lngDetector.addDetector(i18nUserSettingDetector)
-
-  i18next
-    .use(lngDetector)
-    .use(i18nFsBackend)
-    .use(i18nSprintf)
-    .init({
-      // debug: (crowi.node_env === 'development'),
-      fallbackLng: [User.LANG_EN_US],
-      whitelist: Object.keys(User.getLanguageLabels()).map(k => User[k]),
-      backend: {
-        loadPath: crowi.localeDir + '{{lng}}/translation.yml',
-      },
-      detection: {
-        order: ['userSettingDetector', 'header', 'navigator'],
-      },
-      overloadTranslationOptionHandler: i18nSprintf.overloadTranslationOptionHandler,
-    })
+  const middlewares = crowi.middlewares
 
   app.use(function(req, res, next) {
     var now = new Date()
@@ -87,18 +59,7 @@ module.exports = function(crowi, app) {
   app.use(session(crowi.sessionConfig))
 
   // Set basic auth middleware
-  app.use(function(req, res, next) {
-    var config = crowi.getConfig()
-    if (req.query.access_token || req.body.access_token) {
-      return next()
-    }
-
-    if (config.crowi['security:basicName'] && config.crowi['security:basicSecret']) {
-      return basicAuth(config.crowi['security:basicName'], config.crowi['security:basicSecret'])(req, res, next)
-    } else {
-      next()
-    }
-  })
+  app.use(middlewares.BasicAuth)
 
   app.use(passport.initialize())
   app.use(passport.session())
@@ -110,12 +71,6 @@ module.exports = function(crowi, app) {
 
   app.use(middlewares.LoginChecker)
 
-  app.use(i18nMiddleware.handle(i18next))
-  app.use((req, res, next) => {
-    const { _id = '', username: name = '' } = req.user || {}
-    const id = _id.toString()
-    const { language: lang = '' } = req.i18n || {}
-    res.locals.user_context = { user: { id, name }, config: { lang } }
-    next()
-  })
+  app.use(middlewares.I18next)
+  app.use(middlewares.ClientContext)
 }
