+++ lib/middlewares/index.js
@@ -1,11 +1,14 @@
 module.exports = (crowi, app) => ({
   LoginChecker: require('./loginChecker')(crowi, app),
+  BasicAuth: require('./basicAuth')(crowi, app),
+  ClientContext: require('./clientContext')(crowi, app),
   CsrfVerify: require('./csrfVerify')(crowi, app),
   SwigFunctions: require('./swigFunctions')(crowi, app),
   SwigFilters: require('./swigFilters')(crowi, app),
   AdminRequired: require('./adminRequired')(crowi, app),
   LoginRequired: require('./loginRequired')(crowi, app),
   FileAccessRightOrLoginRequired: require('./fileAccessRightOrLoginRequired')(crowi, app),
+  I18next: require('./i18next')(crowi, app),
   AccessTokenParser: require('./accessTokenParser')(crowi, app),
   ApplicationNotInstalled: require('./applicationNotInstalled')(crowi, app),
   ApplicationInstalled: require('./applicationInstalled')(crowi, app),
